# open_api_specification.yaml
---
info:
  title: Virtuális Arborétum API
  version: 0.0.1
  contact:
    name: zslim
    url: https://github.com/zslim
  license:
    name: MIT License
    url: https://mit-license.org/
openapi: 3.0.1
servers:
  - url: http://arboretum-api-dev.herokuapp.com
  - url: http://127.0.0.1:8000
tags:
  - name: Plant
    description: Plant species
  - name: Plant family
    description: Family as in taxonomical rank
    externalDocs:
      description: Learn more on Wikipedia
      url: https://hu.wikipedia.org/wiki/Csal%C3%A1d_(rendszertan)
  - name: Life form
    description: Categories according to Raunkiær's system
    externalDocs:
      description: Learn more on Wikipedia
      url: https://hu.wikipedia.org/wiki/Raunki%C3%A6r-f%C3%A9le_%C3%A9letforma-oszt%C3%A1lyoz%C3%A1s
  - name: Life form subcategory
    description: Subcategories for some life forms, originally used for weed (see Miklós Ujvárosi)
    externalDocs:
      description: Learn more on Wikipedia
      url: https://hu.wikipedia.org/wiki/Raunki%C3%A6r-f%C3%A9le_%C3%A9letforma-oszt%C3%A1lyoz%C3%A1s
paths:
  "/life-forms":
    get:
      summary: Get all life forms
      parameters: [ ]
      responses:
        '200':
          description: Success
      tags:
        - Life form
      description: "Examples: Panerophyta, Hemikryptophyta, Therophyta"
  "/plants":
    get:
      description: get_plants
      parameters: [ ]
      responses:
        '200':
          description: Success
      summary: Get all plants
      tags:
        - Plant
    post:
      description: add_plant
      parameters: [ ]
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/plant"
            examples:
              oxalis:
                "$ref": "#/components/examples/plant_oxalis_corniculata"
      responses:
        '201':
          description: Created
        '400':
          description: Bad request
        '404':
          description: Resource not found
      summary: Add new plant
      tags:
        - Plant
  "/plant-families":
    get:
      summary: Get all plant families
      parameters: [ ]
      responses:
        '200':
          description: Success
      tags:
        - Plant family
    post:
      description: add_plant_family
      parameters: [ ]
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/plant-family"
            examples:
              oxalidaceae:
                "$ref": "#/components/examples/plant_family_oxalidaceae"
      responses:
        '201':
          description: Created
        '400':
          description: Bad request
      summary: Add plant family
      tags:
        - Plant family
  "/plant-families/{family-id}":
    delete:
      summary: Delete plant family
      description: Delete plant family by id
      parameters:
        - in: path
          name: family-id
          required: true
          schema:
            nullable: false
            type: integer
      responses:
        '200':
          description: Success
      tags:
        - Plant family
  "/plants/{plant-id}":
    get:
      summary: Get plant by id
      parameters:
        - in: path
          name: plant-id
          required: true
          schema:
            nullable: false
            type: integer
      responses:
        '200':
          description: Success
      tags:
        - Plant
    delete:
      summary: Delete plant
      description: Delete plant by id
      parameters:
        - in: path
          name: plant-id
          required: true
          schema:
            nullable: false
            type: integer
      responses:
        '200':
          description: Success
      tags:
        - Plant
  "/life-form-subcategories":
    get:
      summary: Get all life form subcategories
      parameters: [ ]
      responses:
        '200':
          description: Success
      tags:
        - Life form subcategory
components:
  schemas:
    plant:
      properties:
        class_taxon:
          nullable: true
          type: string
        decorative_value:
          nullable: true
          type: string
        description_other:
          nullable: true
          type: string
        distance_of_lines:
          nullable: true
          type: string
        distance_of_plants:
          nullable: true
          type: string
        division:
          nullable: true
          type: string
        ecological_function:
          nullable: true
          type: string
        family_id:
          nullable: true
          type: integer
        flower_description:
          nullable: true
          type: string
        frost_tolerance:
          nullable: true
          type: string
        fruit_description:
          nullable: true
          type: string
        gardening_function:
          nullable: true
          type: string
        height:
          nullable: true
          type: string
        humidity_preference:
          nullable: true
          type: string
        hungarian_names:
          nullable: false
          type: array
          items:
            type: string
        id:
          nullable: true
          type: integer
        image_links:
          nullable: true
          type: array
          items:
            type: string
        leaf_description:
          nullable: true
          type: string
        life_form_id:
          nullable: true
          type: integer
        light_preference:
          nullable: true
          type: string
        natural_habitat:
          nullable: true
          type: string
        pests:
          nullable: true
          type: string
        preferred_soil:
          nullable: true
          type: string
        propagation:
          nullable: true
          type: string
        pruning:
          nullable: true
          type: string
        scientific_name:
          nullable: false
          type: string
        time_of_blooming:
          nullable: true
          type: string
        time_of_germination:
          nullable: true
          type: string
        time_of_ripening:
          nullable: true
          type: string
        toxicity:
          nullable: true
          type: string
        usage:
          nullable: true
          type: string
        life_form_subcategory_id:
          nullable: true
          type: integer
      type: object
    plant-family:
      properties:
        hungarian_name:
          nullable: false
          type: string
        id:
          nullable: true
          type: integer
        scientific_name:
          nullable: false
          type: string
      type: object
  examples:
    plant_oxalis_corniculata:
      value:
        scientific_name: Oxalis corniculata
        hungarian_names:
          - szürke madársóska
    plant_family_oxalidaceae:
      value:
        scientific_name: Oxalidaceae
        hungarian_name: madársóskafélék